AdUnit tests were merged into a single `AdUnitSuccessorTests`
